Error fixed in Relion Workflow test due to a randomize subset
The third extraction test (Relion) is to check the correct particles extraction from
a subset of the initial set (i.e. smaller than the CTFs and coords)

This subset was depending to the random subset, thus the assertation procedure with
the first item of the final output sometimes was failed.

Now the subset is done with the splitSet with NO randomize, thus the first item is correct.